requirements.txt
agno
groq
openai
lancedb
tantivy
pypdf
duckduckgo-search
sentence-transformers
python-dotenv
torch